requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
tensorflow==2.15.0
numpy==1.26.4
pillow==12.0.0
python-multipart==0.0.20
